Use more obvious entry save button
