+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Save entries when navigating away
 - Number format for selected measurement in chart
 - Only show dashboard slideshow icon when multiple dashboards are defined
+- Display more obvious entry save button
 
 ## [0.8.116] - 2022-07-26
 ### Fix:
