+++ lib/widgets/journal/entry_details/entry_detail_header.dart
@@ -7,6 +7,7 @@ import 'package:lotti/get_it.dart';
 import 'package:lotti/logic/persistence_logic.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/widgets/journal/entry_datetime_modal.dart';
+import 'package:lotti/widgets/journal/entry_details/save_button.dart';
 import 'package:lotti/widgets/journal/entry_tools.dart';
 import 'package:lotti/widgets/journal/tags_widget.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
@@ -119,6 +120,7 @@ class _EntryDetailHeaderState extends State<EntryDetailHeader> {
                 TagAddIconWidget(itemId: widget.itemId),
               ],
             ),
+            const SaveButton()
           ],
         );
       },
