+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: lotti
 description: A Smart Journal.
 publish_to: 'none'
-version: 0.8.117+1194
+version: 0.8.117+1196
 
 msix_config:
   display_name: Lotti
